requirements.txt
streamlit
langchain
langchain-google-genai
faiss-cpu
tqdm
python-dotenv
tiktoken
langchain-core
langchain-community
langchain_openai
langchain_huggingface